Clippy args change for aleph-node (#772)
* More restrictive clippy checks

* Fixed all clippy complaints

* Fixed clippy args order

* Do not allow type complexity lints to sneak through

* Streamlined clippy command

* First batch

* Partial

* Complex types partial

* Structs for mock data

* Cleanup

* Removed non-channel unused parameters

+++ finality-aleph/src/testing/data_store.rs
@@ -260,7 +260,7 @@ async fn correct_messages_go_through() {
             .await;
 
         for i in 1..=MAX_DATA_BRANCH_LEN {
-            let blocks_branch = blocks[0..(i as usize)].to_vec();
+            let blocks_branch = blocks[0..i].to_vec();
             let test_data: TestData = vec![aleph_data_from_blocks(blocks_branch)];
             test_handler.send_data(test_data.clone());
 
@@ -282,7 +282,7 @@ async fn too_long_branch_message_does_not_go_through() {
 
         test_handler.finalize_block(&blocks[MAX_DATA_BRANCH_LEN + 2].hash());
 
-        let blocks_branch = blocks[0..((MAX_DATA_BRANCH_LEN + 1) as usize)].to_vec();
+        let blocks_branch = blocks[0..(MAX_DATA_BRANCH_LEN + 1)].to_vec();
         let test_data: TestData = vec![aleph_data_from_blocks(blocks_branch)];
         test_handler.send_data(test_data.clone());
         test_handler
@@ -381,7 +381,7 @@ async fn branch_with_not_finalized_ancestor_correctly_handled() {
 
 fn send_proposals_of_each_len(blocks: Vec<Block>, test_handler: &mut TestHandler) {
     for i in 1..=MAX_DATA_BRANCH_LEN {
-        let blocks_branch = blocks[0..(i as usize)].to_vec();
+        let blocks_branch = blocks[0..i].to_vec();
         let test_data: TestData = vec![aleph_data_from_blocks(blocks_branch)];
         test_handler.send_data(test_data.clone());
     }
